refactor: rename Global component to RealEstatePage and update usage in Home; enhance Steps component layout with additional content
