+++ components/RealEstatePage.tsx
